requirements.txt
# Core ML and Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Experiment Tracking
mlflow==2.9.2

# API and Web Framework
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3

# HTTP Requests
requests==2.31.0
urllib3==2.1.0

# DVC and Data Versioning
dvc[azure]==3.30.1
azure-identity==1.15.0
PyYAML
azure-ai-ml==1.12.0
